feat: dedicated retry/update loop mechanism
This commit detaches the update mechanism from the detection. Previously, an
update was only performed after the detection noticed a different ip address.
If the update failed, the program tried again after a short amount of time
but chances were that the retries also failed. After two failed retries, the
program didn't care about the service until the ip address changed again.
That sucked.

With this change we introduce a dedicated "update ticker", which is a loop
that gets executed every second. The loop checks every service for outdated
domains and updates the service if necessary. Services now track their current
number of retries and lock themselves for an increasing amount of time if
subsequent update failures happen.